feat: Phase 4 Implement Allocation core module with complex business rules and tests

+++ src/main/java/com/company/resourceallocation/core/employee/EmployeeService.java
@@ -2,7 +2,9 @@ package com.company.resourceallocation.core.employee;
 
 import com.company.resourceallocation.core.employee.dto.EmployeeRequest;
 import com.company.resourceallocation.core.employee.dto.EmployeeResponse;
+import com.company.resourceallocation.core.allocation.AllocationRepository;
 import com.company.resourceallocation.exception.DuplicateResourceException;
+import com.company.resourceallocation.exception.EmployeeInUseException;
 import com.company.resourceallocation.exception.ResourceNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
@@ -19,6 +21,7 @@ public class EmployeeService {
 
     EmployeeRepository employeeRepository;
     EmployeeMapper employeeMapper;
+    AllocationRepository allocationRepository;
 
     @Transactional
     public EmployeeResponse createEmployee(EmployeeRequest request) {
@@ -69,7 +72,9 @@ public class EmployeeService {
         Employee employee = employeeRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Employee", id));
         
-        // TODO: Phase 4 - check if employee has active allocations before deleting (BR-EMP-03)
+        if (allocationRepository.existsByEmployeeEmployeeId(id)) {
+            throw new EmployeeInUseException("Cannot delete employee: still has active allocations");
+        }
         
         employeeRepository.delete(employee);
     }
